requirements.txt
vsketch==1.1.0
vpype-gcode==0.13.0
pytest==8.3.3
PyHamcrest